Add solution for exercise 4.21 (blog list expansions, step 9)

+++ part_4/blog-list/controllers/blogs.js
@@ -35,6 +35,25 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
+	const blog = await Blog.findById(request.params.id);
+
+	if (blog.user.toString() !== decodedToken.id.toString()) {
+		return response
+			.status(401)
+			.json({
+				error: "deleting a blog is attempted by an invalid user: a blog can be deleted only by the user who added the blog",
+			});
+	}
+
+	const user = await User.findById(decodedToken.id);
+	user.blogs = user.blogs.filter((b) => b.toString() !== blog.id.toString());
+	await user.save();
+
 	await Blog.findByIdAndRemove(request.params.id);
 	response.status(204).end();
 });
